requirements.txt
anaconda-anon-usage @ file:///opt/miniconda3/conda-bld/anaconda-anon-usage_1764636648062/work
anaconda-auth @ file:///opt/miniconda3/conda-bld/anaconda-cloud-auth-split_1765839823423/work
anaconda-cli-base @ file:///opt/miniconda3/conda-bld/anaconda-cli-base_1764888589923/work
annotated-types @ file:///opt/miniconda3/conda-bld/annotated-types_1761744949290/work
anyio==4.12.1
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
boltons @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_97h34kncb9/croot/boltons_1751383736162/work
brotlicffi @ file:///opt/miniconda3/conda-bld/brotlicffi_1764961161081/work
build==1.4.0
certifi @ file:///opt/miniconda3/conda-bld/certifi_1763135329285/work/certifi
cffi @ file:///opt/miniconda3/conda-bld/cffi_1761832773185/work
charset-normalizer @ file:///opt/miniconda3/conda-bld/charset-normalizer_1761744840553/work
chromadb==1.5.2
click @ file:///opt/miniconda3/conda-bld/click_1764332239180/work
conda @ file:///opt/miniconda3/conda-bld/conda_1765567809749/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ebkcs5yxh6/croot/conda-anaconda-tos_1755123330773/work
conda-content-trust @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-content-trust_1728603284288/work
conda-libmamba-solver @ file:///C:/miniconda3/conda-bld/conda-libmamba-solver_1764245615235/work/src
conda-package-handling @ file:///opt/miniconda3/conda-bld/conda-package-handling_1762366402479/work
conda_package_streaming @ file:///opt/miniconda3/conda-bld/conda-package-streaming_1762361629849/work
cryptography @ file:///opt/miniconda3/conda-bld/cryptography-split_1761931964095/work
distro @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/distro_1728591855492/work
durationpy==0.10
faiss-cpu==1.13.2
filelock==3.25.0
flatbuffers==25.12.19
frozendict @ file:///opt/miniconda3/conda-bld/frozendict_1761750665916/work
fsspec==2026.2.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.6.0
idna @ file:///opt/miniconda3/conda-bld/idna_1761912005426/work
importlib_metadata==8.7.1
importlib_resources==6.5.2
jaraco.classes @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_97dr4u6nc1/croot/jaraco.classes_1755516335437/work
jaraco.context @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jaraco.context_1731716684676/work
jaraco.functools @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_91zaq5_008/croot/jaraco.functools_1740408121405/work
Jinja2==3.1.6
joblib==1.5.3
jsonpatch @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jsonpatch_1728594556110/work
jsonpointer @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_bdid71leqg/croot/jsonpointer_1753788445166/work
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring @ file:///opt/miniconda3/conda-bld/keyring_1763637309871/work
kubernetes==35.0.0
libmambapy @ file:///opt/miniconda3/conda-bld/mamba-split_1763111381025/work/libmambapy
markdown-it-py @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6av49xs_3l/croot/markdown-it-py_1756299201304/work
MarkupSafe==3.0.3
mdurl @ file:///opt/miniconda3/conda-bld/mdurl_1758552175942/work
menuinst @ file:///opt/miniconda3/conda-bld/menuinst_1765382369523/work
mmh3==5.2.1
more-itertools @ file:///opt/miniconda3/conda-bld/more-itertools_1761121541654/work
mpmath==1.3.0
msgpack @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0d73gp34a_/croot/msgpack-python_1750958625811/work
networkx==3.6.1
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.3
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079023/work
pkce @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pkce_1728592680499/work
platformdirs @ file:///opt/miniconda3/conda-bld/platformdirs_1762356666077/work
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_bei6q1cvvt/croot/pluggy_1733169621717/work
posthog==5.4.0
protobuf==6.33.5
pybase64==1.4.3
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_61k1o_oje7/croot/pycosat_1736868429741/work
pycparser @ file:///opt/miniconda3/conda-bld/pycparser_1757496048647/work
pydantic @ file:///opt/miniconda3/conda-bld/pydantic_1764083527102/work
pydantic-settings @ file:///opt/miniconda3/conda-bld/pydantic-settings_1764165151428/work
pydantic_core @ file:///opt/miniconda3/conda-bld/pydantic-core_1764009784214/work
Pygments @ file:///opt/miniconda3/conda-bld/pygments_1762431584667/work
PyJWT @ file:///opt/miniconda3/conda-bld/pyjwt_1764332099092/work
PyPika==0.51.1
pyproject_hooks==1.2.0
PySocks @ file:///opt/miniconda3/conda-bld/pysocks_1761753010927/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0bqseadoho/croot/python-dotenv_1745610246418/work
PyYAML==6.0.3
readchar @ file:///opt/miniconda3/conda-bld/readchar_1760613300945/work
referencing==0.37.0
regex==2026.2.28
requests @ file:///opt/miniconda3/conda-bld/requests_1762359592010/work
requests-oauthlib==2.0.0
rich @ file:///opt/miniconda3/conda-bld/rich_1760375480293/work
rpds-py==0.30.0
ruamel.yaml @ file:///opt/miniconda3/conda-bld/ruamel.yaml_1762535951282/work
ruamel.yaml.clib @ file:///opt/miniconda3/conda-bld/ruamel.yaml.clib_1762530070208/work
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semver @ file:///opt/miniconda3/conda-bld/semver_1761903026306/work
sentence-transformers==5.2.3
setuptools==80.9.0
shellingham @ file:///opt/miniconda3/conda-bld/shellingham_1761912204150/work
six==1.17.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03b1up8uh3/croot/tomli_1753774598375/work
torch==2.10.0
tqdm @ file:///opt/miniconda3/conda-bld/tqdm_1762863348900/work
transformers==5.3.0
truststore @ file:///opt/miniconda3/conda-bld/truststore_1762520923820/work
typer==0.20.0
typer-slim==0.20.0
typing-inspection @ file:///opt/miniconda3/conda-bld/typing-inspection_1760614169744/work
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_dffmcwqo6o/croot/typing_extensions_1756281471323/work
urllib3 @ file:///opt/miniconda3/conda-bld/urllib3_1765399201892/work
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wheel==0.45.1
zipp==3.23.0
zstandard @ file:///opt/miniconda3/conda-bld/zstandard_1758189078532/work
